+++ CHANGELOG.md
@@ -11,6 +11,10 @@ incremented for features.
 
 ## [Unreleased]
 
+## Features
+
+* lang: CPI clients for program state instructions.
+
 ## [0.4.2] - 2021-04-10
 
 ## Features