+++ CHANGELOG.md
@@ -13,7 +13,7 @@ incremented for features.
 
 ## Features
 
-* lang: CPI clients for program state instructions.
+* lang: CPI clients for program state instructions ([#43](https://github.com/project-serum/anchor/pull/43)).
 
 ## [0.4.2] - 2021-04-10